Agentic AI Projects/RAG-Based Customer Support Assistant (with LangGraph & HITL)/requirements.txt
langchain==0.3.26
langchain-community==0.3.27
langchain-groq==0.2.4
langchain-google-genai==2.0.10
langgraph==0.2.67
numpy==1.26.4
pypdf==5.0.1
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
python-dotenv==1.0.1
pydantic==2.9.2
pytest==8.3.3
